Provision MOTD on servers
State already in production, as initially prepared for eglide.org:
https://devcentral.nasqueron.org/rOPS6c4c4a16

It deploys a /etc/motd file (or /etc/motd.fail on Debian/Ubuntu),
and a /usr/local/bin/motd command to be able to print it running
`motd`.

We always use /usr/local/bin for executables not deployed as packages.