Grow a forest
**Concept of forest**

The forest is defined as a set of nodes, ie a set of servers.

An example of forest is 'evuaf' for Evuaf nodes, another
one 'woodscloud' for our internal Woods Clouds nodes.

The goal of a forest is to allow to apply a configuration
to different nodes, like populate Evuaf user groups
only on Evuaf servers.

**Pillar**

A new pillar entry called `forests` is provided
to list the available forests.

The first forests defined will be for Evuaf
and Woods Cloud (for our own internal needs).

**Custom execution module**

Starting from this change, complex logic handling
will be stored in custom execution modules.

This change introduces the following Salt method:

forests.exists:

    A function to check if a forest exists.

    CLI Example::

        salt '*' forests.exists evuaf

In a state or a linter, it can be called to check if a
forest value is valid.

**Unit tests**

Salt provides code to test the internal modules in the salt
repository, but not for custom modules.

This change provides mocking capabilities for the dunden
dictionaries like __pillar__.

A makefile allows to run the test suite with `make test`.

**Git configuration**

The new Python code can produce .pyc compiled artefacts.
As such, files are ignored.