Deploy www.idrissi.co content
This change is based on the webserver-content role
from the Nasqueron operations repository.

[ Content ]

Directories are created in the webserver-content role,
with build and deploy instructions.

[ Deploy key ]

As experience showed Salt files copy operations
is a huge slowness bottleneck, it's perhaps preferable
to clone Git repositories on the server instead of using
a staging repository.

As such, we need a deploy key on nodes using webserver-content roles too.